+++ packages/atom.io/src/introspection/attach-transaction-index.ts
@@ -1,0 +1,36 @@
+import type { ReadonlySelectorToken, TransactionToken } from "atom.io"
+import { __INTERNAL__ } from "atom.io"
+
+export const attachTransactionIndex = (
+	store: __INTERNAL__.Store = __INTERNAL__.IMPLICIT.STORE,
+): ReadonlySelectorToken<TransactionToken<unknown>[]> => {
+	const transactionTokenIndexState__INTERNAL = __INTERNAL__.atom__INTERNAL<
+		TransactionToken<unknown>[]
+	>(
+		{
+			key: `👁‍🗨 Transaction Token Index (Internal)`,
+			default: () =>
+				[...store.transactions].map(([key]) => {
+					return { key, type: `transaction` }
+				}),
+			effects: [
+				({ setSelf }) => {
+					store.subject.transactionCreation.subscribe((transactionToken) => {
+						setSelf((state) => [...state, transactionToken])
+					})
+				},
+			],
+		},
+		undefined,
+		store,
+	)
+	const transactionTokenIndex = __INTERNAL__.selector__INTERNAL(
+		{
+			key: `👁‍🗨 Transaction Token Index`,
+			get: ({ get }) => get(transactionTokenIndexState__INTERNAL),
+		},
+		undefined,
+		store,
+	)
+	return transactionTokenIndex
+}
