+++ packages/atom.io/src/introspection/index.ts
@@ -1,0 +1,17 @@
+import type { AtomToken, ReadonlySelectorToken, SelectorToken } from "atom.io"
+
+export * from "./attach-introspection-states"
+
+export type StateTokenIndex<
+	Token extends
+		| AtomToken<unknown>
+		| ReadonlySelectorToken<unknown>
+		| SelectorToken<unknown>,
+> = Record<
+	string,
+	| Token
+	| {
+			key: string
+			familyMembers: Record<string, Token>
+	  }
+>
