--- packages/atom.io/src/react-devtools/meta/index.ts
@@ -1,4 +1,0 @@
-export * from "./attach-meta"
-export * from "./meta-state"
-export * from "./meta-timelines"
-export * from "./meta-transactions"
