--- packages/atom.io/src/react-devtools/meta/meta-timelines.ts
@@ -1,1 +1,0 @@
-export {}
